Secure edit, delete and add product views and profile view

+++ products/views.py
@@ -1,12 +1,11 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
 from .models import Product, Room
 from .forms import ProductForm
-
-# Create your views here.
 
 
 def all_products(request):
@@ -72,14 +71,20 @@ def product_details(request, product_id):
     return render(request, 'products/product_details.html', context)
 
 
+@login_required
 def add_product(request):
     """ Admin add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, staff members only area.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
             product = form.save()
             messages.success(request, 'Successfully added product!')
-            return redirect(reverse('product_details', args=[product.id]))
+            return redirect(reverse(
+                'product_details', args=[product.id]))
         else:
             messages.error(
                 request, 'Failed to add product. Please ensure the form is valid.')
@@ -94,8 +99,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, staff members only area.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -118,8 +128,12 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Admin delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, staff members only area.')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
